lib/mito: move class {Coordinates} outside {geometry} into a basic module that only defines class {Coordinates}
This is necessary to avoid circular headers dependencies. In fact, the {fields} module depends on class {Coordinates}, while the metric objects (which will be soon moved from {manifolds} to {geometry}) depend on the {fields} module.
