+++ lib/mito/coordinates.h
@@ -1,0 +1,18 @@
+// -*- c++ -*-
+//
+// Copyright (c) 2020-2024, the MiTo Authors, all rights reserved
+//
+
+// code guard
+#if !defined(mito_coordinates_h)
+#define mito_coordinates_h
+
+
+// publish the interface
+// the api is in "coordinates/api.h"
+#include "coordinates/public.h"
+
+
+#endif
+
+// end of file
