+++ lib/mito/coordinates/externals.h
@@ -4,16 +4,14 @@
 //
 
 // code guard
-#if !defined(mito_geometry_externals_h)
-#define mito_geometry_externals_h
+#if !defined(mito_coordinates_externals_h)
+#define mito_coordinates_externals_h
 
 
 // externals
 
 // support
-#include "../utilities.h"
-#include "../topology.h"
-#include "../coordinates.h"
+
 
 #endif
