+++ lib/mito/geometry/forward.h
@@ -10,21 +10,6 @@
 
 namespace mito::geometry {
 
-    // available coordinate types
-    enum CoordinateType { CARTESIAN, POLAR, SPHERICAL };
-
-    // class coordinates
-    template <int D, CoordinateType coordT>
-    class Coordinates;
-
-    // concept of a set of coordinates
-    template <class F>
-    concept coordinates_c = requires(F c) {
-        // require that F only binds to {Coordinates} specializations
-        []<int D, CoordinateType coordT>(const Coordinates<D, coordT> &) {
-        }(c);
-    };
-
     // class coordinate system
     template <coordinates_c coordT>
     class CoordinateSystem;
